apply suggestions, make everything work
